+++ HISTORY.md
@@ -67,6 +67,7 @@
 * Blocks, Objects, Threads: new "getImageAttribute" reporter primitive
 * Objects, Threads: let "getImageAttribute" deal with null costumes
 * Objects, Threads: new "stretch" primitive for costumes, also for flipping
+* Threas: new feature: new costume from list of pixels
 
 ### 2019-04-08
 * Blocks, Objects, Threads: new "getSoundAttribute" reporter primitive